Revert "Level 9, Find increment"
