--- data/nova.txt
@@ -1,3 +1,0 @@
-T | 0 | h
-T | 0 | hihihihi
-T | 0 | hihihi
